requirement.txt
alembic==1.13.2
blinker==1.8.2
boto3==1.34.148
botocore==1.34.148
certifi==2024.7.4
click==8.1.7
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
s3transfer==0.10.2
sendgrid==6.11.0
six==1.16.0
SQLAlchemy==2.0.31
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
